Introduce SchedulerBase
