+++ flink-runtime/src/main/java/org/apache/flink/runtime/scheduler/DefaultScheduler.java
@@ -39,7 +39,7 @@ import java.util.concurrent.ScheduledExecutorService;
 /**
  * Stub implementation of the future default scheduler.
  */
-public class DefaultScheduler extends LegacyScheduler {
+public class DefaultScheduler extends SchedulerBase {
 
 	public DefaultScheduler(
 			final Logger log,
